1st commit SETS PRIMARY SCREEN RESOLUTION BASED ON THE DOCKING STATUS
Project derived from A13_EMDK in an attempt to find a workaround to the CHROME CUSTOM TABS rendering issue on the secondary screen. When internal and external DPIs are close enough, the issue disappears.

This version need to run as an app and stay minimized to receive the docking intents. Some issues with sticky broadcast intents at debug time.

The MX feature used (Display / Set screen size) can slightly change the DPIs, not enough to match the 160dpi of the secondary screen.